# docker/swagger.yaml
openapi: 3.0.0
info:
  title: Smart API Hub
  version: 1.0.0
  description: Dynamic REST API platform automatically generated from schema.json.
servers:
  - url: http://localhost:3000
paths:
  /auth/register:
    post:
      summary: Đăng ký người dùng mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                name:
                  type: string
                role:
                  type: string
                  enum: [user, admin]
      responses:
        "201":
          description: Đăng ký thành công
  /auth/login:
    post:
      summary: Đăng nhập nhận JWT Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Đăng nhập thành công, trả về token
...